+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -4,6 +4,7 @@
 - TableCard component: add `onSearch` an `onSort` function props.
 - Add new component `<List />` for displaying interactive list items.
 - Fix z-index issue in `<Chart />` empty message.
+- Added a new `<SimpleSelectControl />` component.
 
 # 3.1.0
 - Added support for a countLabel prop on SearchListItem to allow custom counts.